pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdxl-dora-trainer"
version = "1.0.0"
description = "Production-ready tool for fine-tuning SDXL models using DoRA"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "SDXL DoRA Trainer Contributors"},
]
maintainers = [
    {name = "SDXL DoRA Trainer Contributors"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["stable-diffusion", "sdxl", "dora", "lora", "fine-tuning", "ai", "machine-learning"]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "diffusers>=0.24.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "peft>=0.6.0",
    "safetensors>=0.4.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "rich>=13.0.0",
    "wandb>=0.15.0",
    "tensorboard>=2.14.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=2.0.0",
]
all = [
    "sdxl-dora-trainer[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/yourusername/sdxl-dora-trainer"
Documentation = "https://github.com/yourusername/sdxl-dora-trainer/docs"
Repository = "https://github.com/yourusername/sdxl-dora-trainer"
"Bug Tracker" = "https://github.com/yourusername/sdxl-dora-trainer/issues"
Changelog = "https://github.com/yourusername/sdxl-dora-trainer/blob/main/CHANGELOG.md"

[project.scripts]
sdxl-dora-train = "sdxl_dora_trainer:main"
sdxl-dora-infer = "inference:main"

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*", "examples*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["sdxl_dora_trainer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "--cov=sdxl_dora_trainer",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "setup.py",
    "docs/*",
    "examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
